requirements.txt
streamlit==1.30.0
google-generativeai==0.3.2